+++ src/main/java/com/dataracy/modules/auth/application/dto/response/RegisterTokenResponseDto.java
@@ -1,4 +1,4 @@
-package com.dataracy.modules.user.application.dto;
+package com.dataracy.modules.auth.application.dto.response;
 
 public record RegisterTokenResponseDto(
         String registerToken,
